requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
redis[asyncio]==5.2.1
sqlalchemy==2.0.38
asyncpg==0.30.0
pgvector==0.3.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
sentence-transformers==3.4.1
websockets==14.2
